nfds is an unsigned long

+++ include/nn/socket.h
@@ -26,7 +26,7 @@ using InAddr = ::in_addr;
 Result Initialize(void* pool, ulong poolSize, ulong allocPoolSize, int concurLimit);
 Result Finalize();
 s32 SetSockOpt(s32 socket, s32 socketLevel, s32 option, const void*, u32 len);
-s32 Poll(pollfd *fds, u32 nfds, s32 timeout);
+s32 Poll(pollfd *fds, unsigned long nfds, s32 timeout);
 s32 Fcntl(s32 fd, s32 cmd, ...);
 u64 Send(s32 socket, const void* buffer, u64 bufferLength, s32 flags);
 s32 Socket(s32 domain, s32 type, s32 protocol);
